make ConnectWallet component reuseable
