+++ components/Connect_Wallet/install_metamask.js
@@ -1,0 +1,19 @@
+const InstallMetamask = (params) => {
+  return (
+    <p>
+      {' '}
+      Please install{' '}
+      <a
+        href="https://metamask.io/"
+        className="text-blue-500"
+        rel="noreferrer"
+        target="_blank"
+      >
+        Metamask
+      </a>{' '}
+      crypto wallet.
+    </p>
+  );
+};
+
+export default InstallMetamask;
