+++ components/Navbar/index.js
@@ -1,1 +1,12 @@
-export { default as Navbar } from './main_nav';
+const Navbar = () => {
+  return (
+    <header role="banner">
+      <div>
+        <span>NFT Gram</span>
+        <nav></nav>
+      </div>
+    </header>
+  );
+};
+
+export default Navbar;
