--- components/Navbar/main_nav.jsx
@@ -1,12 +1,0 @@
-const Navbar = () => {
-  return (
-    <header role="banner">
-      <div>
-        <span>NFT Gram</span>
-        <nav></nav>
-      </div>
-    </header>
-  );
-};
-
-export default Navbar;
